refactor(demo-site): split monolithic Demo.tsx into focused modules
Demo.tsx held 74% of src/ (688 lines): 10 components, 4 inline icon SVGs,
URL-state helpers, and all formatting/parsing utilities at one scope. Pull
the reusable and cross-component pieces into their own modules so Demo.tsx
becomes a ~310-line page orchestrator.

- config.ts: CORS proxy, example data, defaults (previously hardcoded in JSX)
- hooks/url.ts: URL-param helpers + new useUrlState(key, fallback) collapsing
  the read-seed-then-debounce-sync pattern
- lib/format.tsx: fmtAmount/safeParseRichText/scaleAmount/formatRichText,
  shared by Demo and Scraper instead of duplicated
- icons.tsx: the 4 inline lucide-style icons
- components/ui.tsx: FOCUS_RING + TextInput, consolidating 4 repeated input
  focus-ring classNames across 3 files
- components/Scraper.tsx: the 235-line scraper, extracted whole

Also: delete unused src/assets/react.svg, rename setURL -> setUrl, dedupe the
two identical nav-button classNames into a NAV_ITEM const.

Pure reorganization — behavior is identical. tsc, eslint, and vite build pass;
verified in-browser (live parse, decomposition, scrape, interactions, no errors).
